refactor: move ChangePreferenceView to morphodict package
